. changed some ugly %lu formating back to %d (caused compile warnings before) . use ircd.conf knock_delay instead of hard-coded 30 seconds, for m_knock
git-svn-id: http://svn.ratbox.org/svnroot/trunk@4223 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/client.c
@@ -1430,8 +1430,8 @@ const char* comment         /* Reason for the exit */
       if (sptr->servptr == &me)
         {
           sendto_realops_flags(FLAGS_ALL,
-		       "%s was connected for %lu seconds.  %d/%d sendK/recvK.",
-			       sptr->name,CurrentTime - sptr->firsttime,
+		       "%s was connected for %d seconds.  %d/%d sendK/recvK.",
+			       sptr->name, (int)(CurrentTime - sptr->firsttime),
 			       sptr->localClient->sendK,
 			       sptr->localClient->receiveK);
           log(L_NOTICE, "%s was connected for %d seconds.  %d/%d sendK/recvK.",
